Add "Accept-ranges: bytes" header for PDF viewing
Adding the "Accept-ranges: bytes" header will allow the pdfjs viewer
to retrieve the PDF in chunks and (when the PDF is web optimized) is
able to display the first pages much faster on large PDF files.
Set PDFJS.disableAutoFetch and PDFJS.disableStream to true in the
pdfjs library viewer to enable fetching only the parts of the PDF file
that are needed for display, so the whole PDF file is not downloaded
if this is not viewed by the user.